Add sparkleSetCanShutdownForUpdate
LiveCode Builder is unable to `send` a message to the LiveCode Script environment during the `CanShutdownForUpdate()` callback. Thus, this library cannot send teh `sparkleShouldPostponeRelaunchForUpdate` message and ask the developer to specify whether or not the update should be installed. This setter allows the developer to set whether or not the app can shutdown.

This is only useful if the application calls this handler whenever the ability to update changes while using the app.